GUI: refactor StdSearchListView to allow passing a custom StdTableSearchList
