+++ src/gui/Src/BasicView/StdSearchListView.h
@@ -2,15 +2,14 @@
 #define STDSEARCHLISTVIEW_H
 
 #include "SearchListView.h"
-#include "StdTable.h"
-
-class StdTableSearchList;
+#include "StdTableSearchList.h"
 
 class StdSearchListView : public SearchListView
 {
     Q_OBJECT
 public:
     StdSearchListView(QWidget* parent, bool enableRegex, bool enableLock);
+    StdSearchListView(QWidget* parent, bool enableRegex, bool enableLock, StdTableSearchList* tableSearchList);
     ~StdSearchListView() override;
 
     void setInternalTitle(const QString & title);
